#add Problem8 #mod return of LogicBase.getResult(Numbers... numbers) set from int to long

+++ ProjectEuler/src/main/java/problems/LogicBase.java
@@ -5,6 +5,6 @@ public abstract class LogicBase {
   public LogicBase() {
   }
 
-  public abstract int getResult(Number... numbers);
+  public abstract long getResult(Number... numbers);
 
 }
